fix(compile-warnings): Fixes a series of compile warnings on MacOS

+++ src/libsyncengine/syncpal/syncpal.cpp
@@ -787,7 +787,7 @@ ExitCode SyncPal::addDlDirectJob(const SyncPath &relativePath, const SyncPath &a
             return;
         }
 
-        if (!setProgress(job->affectedFilePath(), progress)) {
+        if (!setProgress(job->affectedFilePath(), static_cast<int16_t>(progress))) {
             LOGW_SYNCPAL_WARN(_logger, L"Error in SyncPal::setProgress: " << Utility::formatSyncPath(job->affectedFilePath()));
         }
     };
